refactor: move proxy to namespace
