+++ Arbiter.App/App.axaml.cs
@@ -13,7 +13,7 @@ using Arbiter.App.ViewModels.Logging;
 using Arbiter.App.ViewModels.MessageBox;
 using Arbiter.App.ViewModels.Tracing;
 using Arbiter.App.Views;
-using Arbiter.Net;
+using Arbiter.Net.Proxy;
 using Avalonia.Controls;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
