+++ Arbiter.App/ViewModels/Client/ClientViewModel.cs
@@ -1,8 +1,8 @@
 ﻿using System;
-using System.Net.Sockets;
 using Arbiter.Net;
 using Arbiter.Net.Client;
 using Arbiter.Net.Client.Messages;
+using Arbiter.Net.Proxy;
 using Arbiter.Net.Server;
 using Arbiter.Net.Server.Messages;
 using Arbiter.Net.Types;
