+++ Arbiter.App/ViewModels/Tracing/TraceViewModel.cs
@@ -1,21 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.Collections.Specialized;
-using System.ComponentModel;
-using System.IO;
-using System.Linq;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Arbiter.App.Collections;
-using Arbiter.App.Extensions;
 using Arbiter.App.Models;
 using Arbiter.App.Services;
-using Arbiter.Net;
-using Arbiter.Net.Client;
-using Arbiter.Net.Server;
-using Avalonia;
-using Avalonia.Input;
+using Arbiter.Net.Proxy;
 using Avalonia.Platform.Storage;
 using Avalonia.Threading;
 using CommunityToolkit.Mvvm.ComponentModel;
