+++ Arbiter.Net/Proxy/ProxyConnection.RecvLoop.cs
@@ -4,7 +4,7 @@ using Arbiter.Net.Client;
 using Arbiter.Net.Security;
 using Arbiter.Net.Server;
 
-namespace Arbiter.Net;
+namespace Arbiter.Net.Proxy;
 
 public partial class ProxyConnection
 {
