+++ Arbiter.Net/Proxy/ProxyConnection.SendLoop.cs
@@ -3,7 +3,7 @@ using Arbiter.Net.Client;
 using Arbiter.Net.Security;
 using Arbiter.Net.Server;
 
-namespace Arbiter.Net;
+namespace Arbiter.Net.Proxy;
 
 public partial class ProxyConnection
 {
