+++ Arbiter.Net/Proxy/ProxyConnection.cs
@@ -4,7 +4,7 @@ using System.Threading.Channels;
 using Arbiter.Net.Client;
 using Arbiter.Net.Server;
 
-namespace Arbiter.Net;
+namespace Arbiter.Net.Proxy;
 
 public partial class ProxyConnection : IDisposable
 {
