+++ Arbiter.Net/Proxy/ProxyDirection.cs
@@ -1,4 +1,4 @@
-﻿namespace Arbiter.Net;
+﻿namespace Arbiter.Net.Proxy;
 
 public enum ProxyDirection
 {
